@@ -1,6 +1,6 @@
 {
   "name": "dnn-sites-list-view",
-  "version": "0.0.2",
+  "version": "0.0.3",
   "description": "DNN Sites List View",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
@@ -9,24 +9,33 @@
   },
   "license": "UNLICENSED",
   "devDependencies": {
+    "@babel/core": "7.1.6",
+    "@babel/plugin-proposal-object-rest-spread": "7.0.0",
+    "@babel/plugin-transform-object-assign": "7.0.0",
+    "@babel/preset-env": "7.1.6",
+    "@babel/preset-react": "7.0.0",
+    "@dnnsoftware/dnn-react-common": "^2.0.6",
     "babel-loader": "8.0.4",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-react": "6.24.1",
+    "babel-plugin-transform-react-remove-prop-types": "0.4.20",
+    "babel-polyfill": "6.26.0",
+    "create-react-class": "^15.6.3",
     "css-loader": "1.0.1",
-    "dnn-global-styles": "0.0.5",
-    "dnn-webpack-externals": "1.1.2",
-    "eslint": "5.8.0",
-    "eslint-config-dnn": "0.0.12",
+    "eslint": "5.9.0",
     "eslint-import-resolver-webpack": "^0.10.1",
     "eslint-loader": "2.1.1",
-    "eslint-plugin-filenames": "^1.0.0",
+    "eslint-plugin-filenames": "^1.3.2",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-react": "7.11.1",
     "eslint-plugin-spellcheck": "0.0.11",
     "less": "3.8.1",
     "less-loader": "4.1.0",
+    "prop-types": "^15.6.2",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
+    "react-hot-loader": "4.3.5",
     "style-loader": "^0.23.1",
-    "webpack": "4.25.1",
+    "webpack": "4.26.1",
+    "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.10"
   }
 }
