@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import { connect } from "react-redux";
 import {
     siteInfo as SiteInfoActions
@@ -60,7 +61,7 @@ class BasicSettingsPanelBody extends Component {
         }
     }
 
-    componentWillMount() {
+    componentDidMount() {
         const {props} = this;
         if (!this.loadData()) {
             this.setState({
@@ -75,7 +76,7 @@ class BasicSettingsPanelBody extends Component {
         }));
     }
 
-    componentWillReceiveProps(props) {
+    componentDidUpdate(props) {
         let {state} = this;
 
         let title = props.basicSettings["PortalName"];
@@ -194,30 +195,30 @@ class BasicSettingsPanelBody extends Component {
                     <Label
                         tooltipMessage={resx.get("plDescription.Help") }
                         label={resx.get("plDescription") }
-                        />
+                    />
                     <MultiLineInputWithError
                         value={state.basicSettings.Description}
                         onChange={this.onSettingChange.bind(this, "Description") }
                         enabled={canEdit}
-                        />
+                    />
                 </InputGroup>
                 <InputGroup>
                     <Label
                         tooltipMessage={resx.get("plTimeZone.Help") }
                         label={resx.get("plTimeZone") }
-                        />
+                    />
                     <Dropdown
                         options={this.getTimeZoneOptions() }
                         value={state.basicSettings.TimeZone}
                         onSelect={this.onSettingChange.bind(this, "TimeZone") }
                         enabled={canEdit}
-                        />
+                    />
                 </InputGroup>
                 <InputGroup>
                     <Label
                         tooltipMessage={resx.get("plFooterText.Help") }
                         label={resx.get("plFooterText") }
-                        />
+                    />
                     <SingleLineInputWithError
                         inputStyle={{ margin: "0" }}
                         withLabel={false}
@@ -226,7 +227,7 @@ class BasicSettingsPanelBody extends Component {
                         onChange={this.onSettingChange.bind(this, "FooterText") }
                         style={{ width: "100%" }}
                         enabled={canEdit}
-                        />
+                    />
                 </InputGroup>
             </div>;
             const columnTwo = <div className="right-column">
@@ -234,17 +235,18 @@ class BasicSettingsPanelBody extends Component {
                     <Label
                         tooltipMessage={resx.get("plKeyWords.Help") }
                         label={resx.get("plKeyWords") }
-                        />
+                    />
                     <MultiLineInputWithError
                         value={state.basicSettings.KeyWords}
                         onChange={this.onSettingChange.bind(this, "KeyWords") }
-                        enabled={canEdit} />
+                        enabled={canEdit} 
+                    />
                 </InputGroup>
                 <InputGroup>
                     <Label
                         tooltipMessage={resx.get("plGUID.Help") }
                         label={resx.get("plGUID") }
-                        />
+                    />
                     <SingleLineInputWithError
                         inputStyle={{ margin: "0" }}
                         withLabel={false}
@@ -252,13 +254,13 @@ class BasicSettingsPanelBody extends Component {
                         value={state.basicSettings.GUID}
                         style={{ width: "100%" }}
                         enabled={false}
-                        />
+                    />
                 </InputGroup>
                 <InputGroup>
                     <Label
                         tooltipMessage={resx.get("plHomeDirectory.Help") }
                         label={resx.get("plHomeDirectory") }
-                        />
+                    />
                     <SingleLineInputWithError
                         inputStyle={{ margin: "0" }}
                         withLabel={false}
@@ -266,7 +268,7 @@ class BasicSettingsPanelBody extends Component {
                         value={state.basicSettings.HomeDirectory}
                         style={{ width: "100%" }}
                         enabled={false}
-                        />
+                    />
                 </InputGroup>
             </div>;
             const columnThree = <div className="left-column">
@@ -274,7 +276,7 @@ class BasicSettingsPanelBody extends Component {
                     <Label
                         tooltipMessage={resx.get("plLogo.Help") }
                         label={resx.get("plLogo") }
-                        />
+                    />
                     <FileUpload
                         utils={util}
                         portalId={props.portalId}
@@ -282,27 +284,27 @@ class BasicSettingsPanelBody extends Component {
                         folderName={state.basicSettings.LogoFile ? state.basicSettings.LogoFile.FolderName : null}
                         onSelectFile={this.onSettingChange.bind(this, "LogoFile") }
                         fileFormats={["image/png", "image/jpg", "image/jpeg", "image/bmp", "image/gif", "image/jpeg", "image/svg+xml"]}
-                        browseButtonText = {resx.get("BrowseButton")}
-                        uploadButtonText = {resx.get("UploadButton")}
-                        linkButtonText = {resx.get("LinkButton")}
-                        defaultText = {resx.get("DragDefault")}
-                        onDragOverText = {resx.get("DragOver")}
-                        uploadFailedText = {resx.get("UploadFailed")}
-                        wrongFormatText = {resx.get("WrongFormat")}
-                        imageText = {resx.get("DefaultImageTitle")}
-                        linkInputTitleText = {resx.get("LinkInputTitle")}
-                        linkInputPlaceholderText = {resx.get("LinkInputPlaceholder")}
-                        linkInputActionText = {resx.get("LinkInputAction")}
-                        uploadCompleteText = {resx.get("UploadComplete")}
-                        uploadingText = {resx.get("Uploading")}
-                        uploadDefaultText = {resx.get("UploadDefault")}
-                        browseActionText = {resx.get("BrowseAction")}
-                        notSpecifiedText = {resx.get("NotSpecified")}
-                        searchFilesPlaceHolderText = {resx.get("SearchFilesPlaceHolder")}
-                        searchFoldersPlaceHolderText = {resx.get("SearchFoldersPlaceHolder")}
-                        fileText = {resx.get("File")}
-                        folderText = {resx.get("Folder")}
-                        />
+                        browseButtonText={resx.get("BrowseButton")}
+                        uploadButtonText={resx.get("UploadButton")}
+                        linkButtonText={resx.get("LinkButton")}
+                        defaultText={resx.get("DragDefault")}
+                        onDragOverText={resx.get("DragOver")}
+                        uploadFailedText={resx.get("UploadFailed")}
+                        wrongFormatText={resx.get("WrongFormat")}
+                        imageText={resx.get("DefaultImageTitle")}
+                        linkInputTitleText={resx.get("LinkInputTitle")}
+                        linkInputPlaceholderText={resx.get("LinkInputPlaceholder")}
+                        linkInputActionText={resx.get("LinkInputAction")}
+                        uploadCompleteText={resx.get("UploadComplete")}
+                        uploadingText={resx.get("Uploading")}
+                        uploadDefaultText={resx.get("UploadDefault")}
+                        browseActionText={resx.get("BrowseAction")}
+                        notSpecifiedText={resx.get("NotSpecified")}
+                        searchFilesPlaceHolderText={resx.get("SearchFilesPlaceHolder")}
+                        searchFoldersPlaceHolderText={resx.get("SearchFoldersPlaceHolder")}
+                        fileText={resx.get("File")}
+                        folderText={resx.get("Folder")}
+                    />
                 </InputGroup>
             </div>;
             const columnFour = <div className="right-column">
@@ -310,7 +312,7 @@ class BasicSettingsPanelBody extends Component {
                     <Label
                         tooltipMessage={resx.get("plFavIcon.Help") }
                         label={resx.get("plFavIcon") }
-                        />
+                    />
                     <FileUpload
                         utils={util}
                         portalId={props.portalId}
@@ -318,27 +320,27 @@ class BasicSettingsPanelBody extends Component {
                         folderName={state.basicSettings.FavIcon ? state.basicSettings.FavIcon.FolderName : null}
                         onSelectFile={this.onSettingChange.bind(this, "FavIcon") }
                         fileFormats={["image/x-icon", "image/ico"]}
-                        browseButtonText = {resx.get("BrowseButton")}
-                        uploadButtonText = {resx.get("UploadButton")}
-                        linkButtonText = {resx.get("LinkButton")}
-                        defaultText = {resx.get("DragDefault")}
-                        onDragOverText = {resx.get("DragOver")}
-                        uploadFailedText = {resx.get("UploadFailed")}
-                        wrongFormatText = {resx.get("WrongFormat")}
-                        imageText = {resx.get("DefaultImageTitle")}
-                        linkInputTitleText = {resx.get("LinkInputTitle")}
-                        linkInputPlaceholderText = {resx.get("LinkInputPlaceholder")}
-                        linkInputActionText = {resx.get("LinkInputAction")}
-                        uploadCompleteText = {resx.get("UploadComplete")}
-                        uploadingText = {resx.get("Uploading")}
-                        uploadDefaultText = {resx.get("UploadDefault")}
-                        browseActionText = {resx.get("BrowseAction")}
-                        notSpecifiedText = {resx.get("NotSpecified")}
-                        searchFilesPlaceHolderText = {resx.get("SearchFilesPlaceHolder")}
-                        searchFoldersPlaceHolderText = {resx.get("SearchFoldersPlaceHolder")}
-                        fileText = {resx.get("File")}
-                        folderText = {resx.get("Folder")}
-                        />
+                        browseButtonText={resx.get("BrowseButton")}
+                        uploadButtonText={resx.get("UploadButton")}
+                        linkButtonText={resx.get("LinkButton")}
+                        defaultText={resx.get("DragDefault")}
+                        onDragOverText={resx.get("DragOver")}
+                        uploadFailedText={resx.get("UploadFailed")}
+                        wrongFormatText={resx.get("WrongFormat")}
+                        imageText={resx.get("DefaultImageTitle")}
+                        linkInputTitleText={resx.get("LinkInputTitle")}
+                        linkInputPlaceholderText={resx.get("LinkInputPlaceholder")}
+                        linkInputActionText={resx.get("LinkInputAction")}
+                        uploadCompleteText={resx.get("UploadComplete")}
+                        uploadingText={resx.get("Uploading")}
+                        uploadDefaultText={resx.get("UploadDefault")}
+                        browseActionText={resx.get("BrowseAction")}
+                        notSpecifiedText={resx.get("NotSpecified")}
+                        searchFilesPlaceHolderText={resx.get("SearchFilesPlaceHolder")}
+                        searchFoldersPlaceHolderText={resx.get("SearchFoldersPlaceHolder")}
+                        fileText={resx.get("File")}
+                        folderText={resx.get("Folder")}
+                    />
                 </InputGroup>
             </div>;
 
@@ -348,7 +350,7 @@ class BasicSettingsPanelBody extends Component {
                         <Label
                             tooltipMessage={resx.get("plPortalName.Help") }
                             label={resx.get("plPortalName") }
-                            />
+                        />
                         <SingleLineInputWithError
                             inputStyle={{ margin: "0" }}
                             withLabel={false}
@@ -358,27 +360,27 @@ class BasicSettingsPanelBody extends Component {
                             onChange={this.onSettingChange.bind(this, "PortalName") }
                             style={{ width: "100%" }}
                             enabled={canEdit}
-                            />
+                        />
                     </InputGroup>
-                    <Grid children={[columnOne, columnTwo]} numberOfColumns={2} />
+                    <Grid numberOfColumns={2}>{[columnOne, columnTwo]}</Grid>
                     <InputGroup>
                         <Label
                             className={"sectionLabel"}
                             label={resx.get("plLogoIcon") }
-                            />
+                        />
                     </InputGroup>
-                    {canEdit && <Grid children={[columnThree, columnFour]} numberOfColumns={2} />}
+                    {canEdit && <Grid numberOfColumns={2}>{[columnThree, columnFour]}</Grid>}
                     <InputGroup style={{ paddingTop: "10px" }}>
                         <Label
                             tooltipMessage={resx.get("plIconSet.Help") }
                             label={resx.get("plIconSet") }
-                            />
+                        />
                         <Dropdown
                             options={this.getIconSetOptions() }
                             value={state.basicSettings.IconSet}
                             onSelect={this.onSettingChange.bind(this, "IconSets") }
                             enabled={canEdit}
-                            />
+                        />
                     </InputGroup>
                     {canEdit && <div className="buttons-box">
                         <Button
