@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import { connect } from "react-redux";
 import {
     siteBehavior as SiteBehaviorActions
@@ -36,7 +37,7 @@ class SiteAliasesPanel extends Component {
         }
     }
 
-    componentWillMount() {
+    componentDidMount() {
         const {props} = this;
 
         if (!this.loadData()) {
@@ -45,7 +46,7 @@ class SiteAliasesPanel extends Component {
         props.dispatch(SiteBehaviorActions.getSiteAliases(props.portalId));
     }
 
-    componentWillReceiveProps(props) {
+    componentDidUpdate(props) {
         if (tableFields.length === 0) {
             tableFields.push({ "name": resx.get("Alias.Header"), "id": "Alias" });
             tableFields.push({ "name": resx.get("Browser.Header"), "id": "Browser" });
