@@ -14,40 +14,38 @@ const languages = {
 
 module.exports = {
     entry: "./src/main.jsx",
+    optimization: {
+        minimize: isProduction
+    },
     output: {
-        path: "../admin/personaBar/scripts/bundles/",
+        path: path.resolve("../admin/personaBar/scripts/bundles/"),
         filename: "adminLogs-bundle.js",
         publicPath: isProduction ? "" : "http://localhost:8080/dist/"
     },
 
     module: {
-        loaders: [
-            { test: /\.(js|jsx)$/, exclude: /node_modules/, loaders: ["react-hot-loader", "babel-loader"] },
-            { test: /\.less$/, loader: "style-loader!css-loader!less-loader" },
-            { test: /\.(ttf|woff)$/, loader: "url-loader?limit=8192" }
-        ],
-
-        preLoaders: [
-            { test: /\.(js|jsx)$/, exclude: /node_modules/, loader: "eslint-loader" }
+        rules: [
+            { test: /\.(js|jsx)$/, enforce: "pre", exclude: /node_modules/, loader: "eslint-loader", options: { fix: true } },
+            { test: /\.(js|jsx)$/ , exclude: /node_modules/, loader: "babel-loader" },
+            { test: /\.(less|css)$/, loader: "style-loader!css-loader!less-loader" },
+            { test: /\.(ttf|woff)$/, loader: "url-loader?limit=8192" },
         ]
     },
 
     resolve: {
-        extensions: ["", ".js", ".json", ".jsx"],
-        root: [
+        extensions: [".js", ".json", ".jsx"],
+        modules: [
             path.resolve('./src'),          // Look in src first
             path.resolve('./node_modules')  // Last fallback to node_modules
         ]
     },
 
-    externals: require("dnn-webpack-externals"),
+    externals: require("@dnnsoftware/dnn-react-common/WebpackExternals"),
 
     plugins:
     isProduction
         ?
         [
-            new webpack.optimize.UglifyJsPlugin(),
-            new webpack.optimize.DedupePlugin(),
             new webpack.DefinePlugin({
                 VERSION: JSON.stringify(packageJson.version),
                 "process.env": {
