@@ -64,7 +64,6 @@ module.exports = {
       "space-before-blocks": "warn",
       "no-multiple-empty-lines":  "warn",
       "react/jsx-equals-spacing": ["warn", "never"],
-      "id-match": ["error", "^([A-Za-z0-9_])+$", {"properties": true}],
-      "no-useless-escape": "off"
+      "id-match": ["error", "^([A-Za-z0-9_])+$", {"properties": true}]
     }
 };
