@@ -1,4 +1,5 @@
-import React, {PropTypes} from "react";
+import PropTypes from 'prop-types';
+import React from "react";
 
 /* eslint-disable react/no-danger */
 const Statistic = ({label, value}) => (
