@@ -2,7 +2,7 @@
 const path = require("path");
 const packageJson = require("./package.json");
 const isProduction = process.env.NODE_ENV === "production";
-const webpackExternals = require("dnn-webpack-externals");
+const webpackExternals =  require("@dnnsoftware/dnn-react-common/WebpackExternals");
 const languages = {
     "en": null
     // TODO: create locallizaton files per language 
@@ -15,31 +15,30 @@ const languages = {
 const moduleName = "sites";
 module.exports = {
     entry: "./src/main.jsx",
+    optimization: {
+        minimize: isProduction
+    },
     output: {
-        path: "../admin/personaBar/scripts/bundles/",
+        path: path.resolve("../admin/personaBar/scripts/bundles/"),
         filename: moduleName + "-bundle.js",
         publicPath: isProduction ? "" : "http://localhost:8080/dist/"
     },
 
     module: {
-        loaders: [
-            { test: /\.(js|jsx)$/, exclude: /node_modules/, loaders: ["react-hot-loader", "babel-loader"] },
-            { test: /\.less$/, loader: "style-loader!css-loader!less-loader" },
+        rules: [
+            { test: /\.(js|jsx)$/, enforce: "pre", exclude: /node_modules/, loader: "eslint-loader", options: { fix: true } },
+            { test: /\.(js|jsx)$/ , exclude: /node_modules/, loader: "babel-loader" },
+            { test: /\.(less|css)$/, loader: "style-loader!css-loader!less-loader" },
             { test: /\.(ttf|woff)$/, loader: "url-loader?limit=8192" }
-        ],
-
-        preLoaders: [
-            { test: /\.(js|jsx)$/, exclude: /node_modules/, loader: "eslint-loader" }
         ]
     },
-
     resolve: {
-        extensions: ["", ".js", ".json", ".jsx"],
-        root: [
+        extensions: [".js", ".json", ".jsx"],
+        modules: [
             path.resolve('./src'),          // Look in src first
             path.resolve('./exportables'),  // Look in exportables after
             path.resolve('./node_modules')  // Last fallback to node_modules
-        ]
+        ]        
     },
     externals: Object.assign(webpackExternals, {
         "dnn-sites-common-action-types": "window.dnn.Sites.CommonActionTypes",
@@ -48,8 +47,6 @@ module.exports = {
         "dnn-sites-common-actions": "window.dnn.Sites.CommonActions"
     }),
     plugins: isProduction ? [
-        new webpack.optimize.UglifyJsPlugin(),
-        new webpack.optimize.DedupePlugin(),
         new webpack.DefinePlugin({
             VERSION: JSON.stringify(packageJson.version),
             "process.env": {
