@@ -22,11 +22,11 @@ module.exports = {
         "VERSION": false
     },
     "rules": {
-       "spellcheck/spell-checker": [1,
+       /*"spellcheck/spell-checker": [1,
         {
-            "comments": "true",
-            "strings": "true",
-            "identifiers": "true",
+            "comments": true,
+            "strings": true,
+            "identifiers": true,
             "skipWords": require("./.eslintskipwords"),
             "skipIfMatch": [
                 "http://[^s]*",
@@ -34,7 +34,7 @@ module.exports = {
                 "(^#[0-9A-F]{6}$)|(^#[0-9A-F]{3}$)" // CSS hex color
             ]
         }
-      ],      
+       ],*/     
       "semi": "error",
       "no-var": "error",
       "quotes": ["warn", "double" ],
