@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import Collapse from "dnn-collapsible";
 import { FolderIcon, PagesIcon, ArrowRightIcon, ArrowDownIcon } from "dnn-svg-icons";
 class Folder extends Component {
