@@ -1,4 +1,5 @@
-import React, { PropTypes, Component } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import resx from "resources";
 import Switch from "dnn-switch";
 import Button from "dnn-button";
