@@ -1,9 +1,8 @@
-import React, { Component, PropTypes } from "react";
-import Modal from "dnn-modal";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import Button from "dnn-button";
 import TextOverflowWrapper from "dnn-text-overflow-wrapper";
 import resx from "resources";
-import utilities from "utils";
 import "./fullEditor.less";
 
 class FullEditor extends Component {
@@ -13,94 +12,98 @@ class FullEditor extends Component {
         };
 
         this.editorToolbar = [
-            ['undo', 'redo'],
-            ['bold', 'italic', 'underline', 'strikeThrough'],
-            ['justifyLeft', 'justifyCenter', 'justifyRight', 'justifyFull'],
-            ['insertUnorderedList', 'insertOrderedList']
+            ["undo", "redo"],
+            ["bold", "italic", "underline", "strikeThrough"],
+            ["justifyLeft", "justifyCenter", "justifyRight", "justifyFull"],
+            ["insertUnorderedList", "insertOrderedList"]
         ];
     }
 
-    componentDidMount(){
+    componentDidMount() {
         this.editorControl.focus();
     }
 
-    execCommand(button, e){
+    execCommand(button, e) {
         this.editorControl.focus();
         document.execCommand(button, false, null);
         
         e.preventDefault();
     }
 
-    onSave(){
+    onSave() {
         const { props } = this;
 
-        if(typeof props.onChange === "function"){
+        if (typeof props.onChange === "function") {
             let content = this.editorControl.innerHTML;
             props.onChange(content);
         }
     }
 
-    onCancel(){
+    onCancel() {
         const { props } = this;
 
         props.onCancel();
     }
 
     /* eslint-disable react/no-danger */
-    renderToolbar(){
-        const { props } = this;
-
-        return <div className='fulleditor-controls'>
-        {
-            this.editorToolbar.map((group) => {
-              return <div className='btn-group'>
-                  {
-                      group.map((button) => {
-                          return <a 
-                                className='btn' 
-                                data-role={button} 
-                                href='#'
-                                onClick={this.execCommand.bind(this, button)}
-                                dangerouslySetInnerHTML={{ __html: require('./icons/' + button + '.svg') }}
-                            >
-                              </a>;
-                      })
-                  }
-              </div>;
-            })
-        }
-        </div>;
+    renderToolbar() {
+        return (
+            <div className='fulleditor-controls'>
+                {
+                    this.editorToolbar.map((group, i) => {
+                        return <div className='btn-group' key={i}>
+                            {
+                                group.map((button, i) => {
+                                    return (
+                                        <a 
+                                            className='btn' 
+                                            data-role={button} 
+                                            href='#'
+                                            onClick={this.execCommand.bind(this, button)}
+                                            dangerouslySetInnerHTML={{ __html: require("./icons/" + button + ".svg") }}
+                                            key={i}>
+                                        </a>
+                                    );
+                                })
+                            }
+                        </div>;
+                    })
+                }
+            </div>
+        );
     }
 
     /* eslint-disable react/no-danger */
-    renderEditor(){
+    renderEditor() {
         const { props } = this;
-        return <div className='fulleditor-editor' 
-                    contentEditable
-                    ref={(e) => { this.editorControl = e; }}
-                    dangerouslySetInnerHTML={{ __html: props.value }}>
-                </div>;
+        return (
+            <div className='fulleditor-editor' 
+                contentEditable
+                ref={(e) => { this.editorControl = e; }}
+                dangerouslySetInnerHTML={{ __html: props.value }}>
+            </div>
+        );
     }
 
-    renderButtons(){
+    renderButtons() {
         return [
-            <Button type="secondary" onClick={this.onCancel.bind(this)}>
+            <Button type="secondary" onClick={this.onCancel.bind(this)} key="first">
                 <TextOverflowWrapper text={resx.get("Cancel") } maxWidth={100} />
             </Button>,
-            <Button type="primary" onClick={this.onSave.bind(this)}>
+            <Button type="primary" onClick={this.onSave.bind(this)} key="second">
                 <TextOverflowWrapper text={resx.get("Save") } maxWidth={100} />
             </Button>
         ];
     }
 
     render() {
-        const { props } = this;
-
-        return (<div className="dnn-language-resource-full-editor">
-          {this.renderToolbar()}
-          {this.renderEditor()}
-          {this.renderButtons()}
-            </div>);
+        return (
+            <div className="dnn-language-resource-full-editor">
+                {this.renderToolbar()}
+                {this.renderEditor()}
+                {this.renderButtons()}
+            </div>
+        );
     }
 }
 
