@@ -19,39 +19,77 @@ const webpackExternals = Object.assign({},
 
 module.exports = {
     entry: "./src/main.jsx",
-    output: {
-        path: "../admin/personaBar/scripts/bundles/",
-        filename: "site-settings-bundle.js",
-        publicPath: isProduction ? "" : "http://localhost:8085/dist/"
+    optimization: {
+        minimize: isProduction
     },
-
-    module: {
-        loaders: [
-            { test: /\.(js|jsx)$/, exclude: /node_modules/, loaders: ["react-hot-loader", "babel-loader"] },
-            { test: /\.less$/, loader: "style-loader!css-loader!less-loader" },
-            { test: /\.(ttf|woff)$/, loader: "url-loader?limit=8192" },
-            { test: /\.(svg)$/, exclude: /node_modules/, loader: "raw-loader" }
-        ],
-
-        preLoaders: [
-            { test: /\.(js|jsx)$/, exclude: /node_modules/, loader: "eslint-loader" },
-            { test: /\.js$/, loader: "source-map-loader" }
-        ]
+    output: {
+        path: path.resolve(__dirname, '../admin/personaBar/scripts/bundles/'),
+        publicPath: isProduction ? "" : "http://localhost:8085/dist/",
+        filename: "site-settings-bundle.js"
     },
 
     resolve: {
-        extensions: ["", ".js", ".json", ".jsx"],
-        root: [
-            path.resolve('./src'),          // Look in src first
-            path.resolve('./node_modules')  // Last fallback to node_modules
+        extensions: ["*", ".js", ".json", ".jsx"],
+        modules: [
+            path.resolve('./src'),           // Look in src first
+            path.resolve('./node_modules')   // Last fallback to node_modules
         ]
     },
-
+    module: {
+        rules: [
+            { 
+                test: /\.(js|jsx)$/, 
+                exclude: /node_modules/, 
+                enforce: "pre",
+                use: [
+                    'eslint-loader'
+                ] 
+            },
+            { 
+                test: /\.js$/, 
+                enforce: "pre",
+                use: [
+                    'source-map-loader'
+                ] 
+            },
+            { 
+                test: /\.less$/, 
+                use: [{
+                    loader: 'style-loader'  // creates style nodes from JS strings
+                }, {
+                    loader: 'css-loader'    // translates CSS into CommonJS
+                }, {
+                    loader: 'less-loader'   // compiles Less to CSS
+                }]
+            },
+            { 
+                test: /\.(js|jsx)$/, 
+                exclude: /node_modules/, 
+                use: {
+                    loader: 'babel-loader',
+                    options: {
+                        presets: ['@babel/preset-env','@babel/preset-react']
+                    }
+                } 
+            },
+            {
+                test: /\.(ttf|woff)$/, 
+                use: {
+                    loader: 'url-loader?limit=8192'
+                }
+            },
+            { 
+                test: /\.(svg)$/, 
+                exclude: /node_modules/, 
+                use: [
+                    'raw-loader'
+                ] 
+            }
+        ]
+    },
     externals: webpackExternals,
 
     plugins: isProduction ? [
-        new webpack.optimize.UglifyJsPlugin(),
-        new webpack.optimize.DedupePlugin(),
         new webpack.DefinePlugin({
             VERSION: JSON.stringify(packageJson.version),
             "process.env": {
